+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/ryansarll/Desktop/cat-tinder-frontend/src/App.js":"1"},{"size":1054,"mtime":1609963105566,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"5i880i","/Users/ryansarll/Desktop/cat-tinder-frontend/src/App.js",[]]
+[{"/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/App.js":"1","/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/index.js":"2","/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/mockCats.js":"3","/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/reportWebVitals.js":"4","/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/pages/CatNew.js":"5","/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/pages/NotFound.js":"6","/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/components/Footer.js":"7","/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/components/Header.js":"8"},{"size":1181,"mtime":1609972723697,"results":"9","hashOfConfig":"10"},{"size":547,"mtime":1609972964306,"results":"11","hashOfConfig":"10"},{"size":342,"mtime":1609972242412,"results":"12","hashOfConfig":"10"},{"size":362,"mtime":1609971350735,"results":"13","hashOfConfig":"10"},{"size":210,"mtime":1609972593522,"results":"14","hashOfConfig":"10"},{"size":230,"mtime":1609972593522,"results":"15","hashOfConfig":"10"},{"size":210,"mtime":1609974269678,"results":"16","hashOfConfig":"10"},{"size":544,"mtime":1609974511861,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"tdg8lj",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"23"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"30","messages":"31","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/App.js",[],["36","37"],"/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/index.js",[],["38","39"],"/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/mockCats.js",[],"/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/reportWebVitals.js",[],"/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/pages/CatNew.js",[],"/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/pages/NotFound.js",[],"/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/components/Footer.js",[],"/Users/learnacademy/Desktop/cat-tinder-frontend-chris-ryan/src/components/Header.js",[],{"ruleId":"40","replacedBy":"41"},{"ruleId":"42","replacedBy":"43"},{"ruleId":"40","replacedBy":"44"},{"ruleId":"42","replacedBy":"45"},"no-native-reassign",["46"],"no-negated-in-lhs",["47"],["46"],["47"],"no-global-assign","no-unsafe-negation"]
