--- components/Header.js
@@ -1,20 +1,0 @@
-import React, { Component } from 'react'
-import { Jumbotron, Container } from 'reactstrap'
-
-class Header extends Component {
-    render(){
-        return(
-            <>
-                <h1>Header</h1>
-                <Jumbotron fluid>
-                    <Container fluid>
-                        <h1 className="display-3">Fluid jumbotron</h1>
-                        <p className="lead">This is a modified jumbotron that occupies the entire horizontal space of its parent</p>
-                    </Container>
-                </Jumbotron>
-            </>
-        )
-    }
-}
-
-export default Header
