+++ src/App.test.js
@@ -2,7 +2,10 @@ import { render, screen } from '@testing-library/react';
 import App from './App';
 
 test('renders learn react link', () => {
+  
   render(<App />);
+  
   const linkElement = screen.getByText(/learn react/i);
+  
   expect(linkElement).toBeInTheDocument();
 });
