+++ src/components/Footer.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react'
+
+class Footer extends Component {
+    render(){
+        return(
+            <>
+                <h1>Footer</h1>
+
+            </>
+        )
+    }
+}
+
+
+export default Footer
