+++ src/components/Header.js
@@ -1,15 +1,15 @@
 import React, { Component } from 'react'
 import { Jumbotron, Container } from 'reactstrap'
 
-class Footer extends Component {
+class Header extends Component {
     render(){
         return(
             <>
-                <h1>Footer</h1>
+                <h1>Header</h1>
                 <Jumbotron fluid>
                     <Container fluid>
-                        <h1 className="display-3">Fluid jumbotron</h1>
-                        <p className="lead">This is a modified jumbotron that occupies the entire horizontal space of its parent</p>
+                        <h1 className="display-3">FelineFarmers.com</h1>
+                        <p className="lead">City Kitties just don't get it</p>
                     </Container>
                 </Jumbotron>
             </>
@@ -17,5 +17,4 @@ class Footer extends Component {
     }
 }
 
-
-export default Footer
+export default Header
