+++ src/mockCats.js
@@ -1,0 +1,18 @@
+let catz = [
+    {   id: 1, 
+        name:"Cody",
+        age: 4,
+        enjoys:"playing with strings"
+     },
+     {  id: 2, 
+        name:"Cindy",
+        age: 5,
+        enjoys:"hanging out in the barn"
+     },
+     {  id: 3, 
+        name:"Jack",
+        age: 3,
+        enjoys:"spending time with the cows"
+     },
+]
+export default catz
