+++ src/pages/NotFound.js
